+++ src/python/layouts.py
@@ -1,0 +1,28 @@
+import PySimpleGUI as sg
+sg.set_options(font=('Franklin Gothic Book', 24))
+b_style1 = {'size':(5,2),'button_color':("#00b894","#dfe6e9")}
+b_style2 = {'size':(5,2),'button_color':("white","#0984e3")}
+b_style3 = {'size':(5,2),'button_color':("white","tomato")}
+b_style4 = {'size':(9,2),'button_color':("#00b894","#dfe6e9")}
+font = ("Arial", 13)
+
+layout_calculator=[
+    [sg.Button('s'),sg.Text('Calculator', size=(30,1), justification = 'right',background_color='#909090')],
+    [sg.Text('0.00000', key='output', size=(30,1), justification = 'right', background_color='white', text_color='black',pad=(5,0))],
+    [sg.Text('0.00000', key='current_output', size=(63,1), justification = 'right', background_color='#909090', text_color='black',pad=(5,0),font=font)],
+    [sg.Button("7",**b_style1),sg.Button("8",**b_style1),sg.Button("9",**b_style1),sg.Button("DEL",**b_style2),sg.Button("AC",**b_style2)],
+    [sg.Button("4",**b_style1),sg.Button("5",**b_style1),sg.Button("6",**b_style1),sg.Button("x",**b_style1),sg.Button("/",**b_style1)],
+    [sg.Button("1",**b_style1),sg.Button("2",**b_style1),sg.Button("3",**b_style1),sg.Button("+",**b_style1),sg.Button("-",**b_style1),],
+    [sg.Button("0",**b_style1),sg.Button(".",**b_style1),sg.Button("=",**b_style1),sg.Button("^",**b_style1),sg.Button("OFF",**b_style3),],
+    [sg.Button("(",**b_style4),sg.Button(")",**b_style4),sg.Button("√", key="root",**b_style4)]
+    ]
+
+layout_settings=[
+    [sg.Checkbox("Dark mode",background_color='#909090')]
+]
+
+currentLayout=0
+
+layout=[[sg.Column(layout_calculator,visible=True, key='LayoutC',background_color='#909090'), sg.Column(layout_settings,visible=False, key="LayoutS",background_color='#909090')]]
+
+
